bmi calculator change 2

+++ demogit/Day 2/Day2.2-BMI_Calculator.py
@@ -4,6 +4,6 @@ weight = input("what is your weight in kg: ")
 a = int(weight)
 b = float(height)
 
-bmi = a / (b ** 2)
+bmi = int(a / (b ** 2))
 
-print(int(bmi))
+print(bmi)
